Update serializers to current versions (10-03-2019).
That means that the updated versions are now:

* thrift 0.12.0
* protobuf 3.7.0
* boost 1.69.0
* msgpack 3.1.1
* cereal 1.2.2
* avro 1.8.2
* capnproto 0.7.0
* flatbuffers 1.10.0
* YAS 7.0.2

Only avro was not updated.

This version of capnproto requires C++14.  This change in requirements
is reflected in CMakeLists.txt and README.md.

The benchmark is performed on my system, which has slightly different
specs from the original author's.  This is also reflected in the
README.md.  Graphs are updated to reflect these results.